Extract glyph renderer to printer_glyph.cpp
Moves the typography / rasterisation code out of the main printer.cpp
TU:

- Printer::FillPalette  (palette setup, 8 sub-palettes of 32 colours)
- Printer::UpdateFont   (FreeType face loading, size + transform)
- Printer::BlitGlyph    (copy FT_Bitmap onto the page surface)
- Printer::DrawLine     (under/strike/overscore lines)

These four methods are the bulk of the 'how a glyph ends up on the
page' machinery; collecting them in one file makes the dispatch in
printer.cpp easier to follow.

Also drops a stale /* switch(style & (STYLE_CONDENSED|...)) */
comment block that documented an upstream alternative code path.
Wired into src/hardware/CMakeLists.txt.
